corrigindo phpdocs e criando testes
