+++ tests/BootstrapTrait.php
@@ -2,17 +2,16 @@
 
 namespace PhpSigep\Test;
 
+use PhpSigep\Config;
+
 trait BootstrapTrait
 {
     public function setUpSIGEP(): void
     {
         $config = new \PhpSigep\Config();
-
         $config->setAccessData(new \PhpSigep\Model\AccessDataHomologacao());
-
         $config->setEnv(\PhpSigep\Config::ENV_DEVELOPMENT);
-
-
+        $config->setXmlEncode(Config::XML_ENCODE_UTF);
         $config->setCacheOptions(array(
             'storageOptions' => array(
                 'enabled' => false,
